scraper/requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
psycopg[binary]==3.2.10
requests==2.32.3
sentry-sdk==2.19.0
